Add files via upload
Python Version of the Project.

The objective of this project was to minimize the mean squared error,  reduce multicollinearity, and simplify our model. 

This notebook is the results of investigating the dataset housing.proper.csv. 

Of 13 variables, we will keep 7 and discard 6.  We will also apply the log transformation on the desired output.

The objective of this project was to minimize the mean squared error,  reduce multicollinearity, and simplify our model.